requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
psycopg[binary]>=3.1.19
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.32.0
eth-utils>=4.1.1
eth-hash[pycryptodome]>=0.8.0
email-validator>=2.2.0
